Fix bug in pyx/visit-test-module.
When working on a namespace package the function has problems locating
the "tests" directory. If the package is called "foo.bar" it assumes
"foo/tests" not "foo/bar/tests".

The new implementation simply searches the "tests" directory from the
current directory, upwards until it's found or the root of the package
is reached.